+++ flockoff/miners/model.py
