+++ pyproject.toml
@@ -5,7 +5,7 @@ description = "FLock Subnet"
 authors = ["Flock Devs"]
 readme = "README.md"
 packages = [
-    {include = "FLockDataset"},
+    {include = "flockoff"},
     {include = "neurons"}
 ]
